backend/requirements.txt
Django>=5.0,<6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
Pillow==10.4.0
python-dotenv==1.0.1
cloudinary>=1.41
# Production database
dj-database-url==2.2.0
psycopg2-binary==2.9.9
# Optional S3 storage
boto3==1.35.1
django-storages==1.14.4
# --- Added for semantic embeddings ---
sentence-transformers==3.0.1
numpy==2.1.3
# --- Added for Google auth ---
google-auth==2.34.0
# Production server
gunicorn==21.2.0

# ============================================
# RAG & Vector Database
# ============================================
pinecone-client==3.0.0
langchain==0.1.10
langchain-openai==0.0.11
langchain-community==0.0.20
openai==1.3.8
tiktoken==0.5.2
PyPDF2==3.0.1
python-docx==0.8.11

# ============================================
# WhatsApp & Messaging
# ============================================
requests==2.31.0
mailjet-rest==1.3.4

# ============================================
# Text Processing & NLP
# ============================================
nltk==3.8.1
textblob==0.17.1
spacy==3.7.2

# ============================================
# Async & Queuing
# ============================================
celery==5.3.4
redis==5.0.1
python-json-logger==2.0.7

# ============================================
# Validation & Data Processing
# ============================================
pydantic==2.5.0
pydantic-settings==2.1.0

# ============================================
# Monitoring & Logging
# ============================================
sentry-sdk==1.38.0

# ============================================
# Testing
# ============================================
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
factory-boy==3.3.0

# ============================================
# Machine Learning & AI
# ============================================
scikit-learn==1.3.2
pandas==2.1.4
joblib==1.3.2
xgboost==2.0.3
